refactor: update ColorDisplay component to use <script setup> syntax, enhance FeedbackDialog layout, and introduce CardProduct component for improved product display in the application
