+++ src/pages/ProductsPage.vue
@@ -43,6 +43,8 @@
             round
             size="sm"
           />
+
+          <!-- Create Product -->
           <q-btn
             icon="mdi-plus"
             round
